🚚 Extract mock-identity-provider-fca-low to standalone app
**Problem**

mock-identity-provider-fca-low lived inside the back/ NestJS monorepo
as a plain Express + oidc-provider app with zero NestJS runtime
imports. It dragged in the full nest build toolchain, back/Dockerfile's
6-stage build, and back/package.json's ~80 mostly-irrelevant deps just
to run a static OIDC mock.

**Proposal**

Extract it to a root-level mock-identity-provider/ directory running
plain Node 24 with --experimental-strip-types (no build step, no
NestJS). Add /livez and /readyz health endpoints for Docker health
checks and compose dependency conditions, migrate to lodash-es, and
update CI (build-docker.yml), dependabot, and compose (mocks.yml,
shared/compose.yaml, fca-low env files) to point at the new location.
