+++ .github/workflows/build-docker.yml
@@ -96,7 +96,7 @@ jobs:
           - core-fca-low-migrator
           - csmr-rie
           - mock-data-provider
-          - mock-identity-provider-fca-low
+          - mock-identity-provider
           - mock-service-provider-fca-low
           - api-partner
           - fc-exploitation-v2
@@ -110,8 +110,8 @@ jobs:
             context: ./back
           - instance: mock-data-provider
             context: ./mock-data-provider
-          - instance: mock-identity-provider-fca-low
-            context: ./back
+          - instance: mock-identity-provider
+            context: ./mock-identity-provider
           - instance: mock-service-provider-fca-low
             context: ./back
           - instance: api-partner
