+++ docker/compose/fca-low/.env/moncomptepro.env
@@ -1,11 +1,8 @@
 APP_NAME=moncomptepro-low
-NESTJS_INSTANCE=mock-identity-provider-fca-low
 FQDN=moncomptepro.docker.dev-franceconnect.fr
 REQUEST_TIMEOUT=6000
 App_CITIZEN_DATABASE_PATH=/var/databases/citizen
 App_SCENARIOS_DATABASE_PATH=/var/databases/scenarios/scenarios.json
-App_ASSETS_PATHS=["../../../apps/mock-identity-provider-fca-low/src"]
-App_VIEWS_PATHS=["../../../apps/mock-identity-provider-fca-low//src"]
 App_ALLOW_CUSTOM_IDENTITY=true
 
 ## @todo Replace `html` by `json` for the React app to receive error message in JSON format instead of a HTML view.
