+++ docker/compose/shared/compose.yaml
@@ -120,22 +120,27 @@ services:
       - public
       - data
 
-  base-idp-mock:
-    extends: base-mock
-    volumes:
-      - "../../volumes/app:/opt/scripts"
-      - "../../volumes/fca-low/mocks/idp/databases:/var/databases"
-      - "../../volumes/log/:/var/log/app"
-      - "../../volumes/.home:/home"
-      - "../../volumes/ssl:/etc/ssl/docker_host:ro"
+  base-mock-identity-provider:
+    build:
+      context: "../../../mock-identity-provider"
+      target: dev
+    user: ${CURRENT_UID}
     depends_on:
       rp-all:
         condition: service_started
-      redis-pwd:
-        condition: service_healthy
+    healthcheck:
+      test:
+        - "CMD"
+        - "wget"
+        - "-qO"
+        - "/dev/null"
+        - "http://localhost:3000/readyz"
+    tty: true
+    init: true
+    volumes:
+      - "../../../mock-identity-provider/src:/var/www/app/src:ro"
     networks:
       - public
-      - data
 
   base-mock-data-provider:
     build:
