+++ mock-identity-provider/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "mock-identity-provider",
+  "version": "1.0.0",
+  "type": "module",
+  "scripts": {
+    "start": "node src/main.ts",
+    "start:dev": "node --watch src/main.ts",
+    "lint": "tsc --noEmit && prettier --check .",
+    "format": "prettier --write .",
+    "test": "node --test src/app.test.ts"
+  },
+  "dependencies": {
+    "ejs": "^6.0.1",
+    "express": "^5.2.1",
+    "lodash-es": "^4.18.1",
+    "oidc-provider": "^8.8.0",
+    "quick-lru": "^7.3.0",
+    "zod": "^4.4.3"
+  },
+  "devDependencies": {
+    "@tsconfig/strictest": "^2.0.8",
+    "@types/ejs": "^3.1.5",
+    "@types/express": "^5.0.6",
+    "@types/lodash-es": "^4.17.12",
+    "@types/node": "^22.5.5",
+    "@types/oidc-provider": "8.8.1",
+    "@types/supertest": "^7.2.1",
+    "prettier": "^3.9.6",
+    "prettier-plugin-organize-imports": "^4.3.0",
+    "supertest": "^7.2.2",
+    "typescript": "^7.0.2"
+  },
+  "prettier": {
+    "plugins": [
+      "prettier-plugin-organize-imports"
+    ]
+  }
+}
