+++ mock-identity-provider/src/oidc-provider-support/policy.ts
@@ -1,10 +1,12 @@
-import { get, isEmpty } from "lodash";
+import { get, isEmpty } from "lodash-es";
 import { interactionPolicy } from "oidc-provider";
 
 const { Check, base } = interactionPolicy;
 const basePolicy = base();
 
 const loginPrompt = basePolicy.get("login");
+if (!loginPrompt) throw new Error("base policy missing login prompt");
+
 loginPrompt.checks.remove("essential_acr");
 loginPrompt.checks.add(
   new Check(
@@ -12,7 +14,10 @@ loginPrompt.checks.add(
     "requested ACR could not be obtained",
     (ctx) => {
       const { oidc } = ctx;
-      const request = get(oidc.claims, "id_token.acr", {});
+      const request = get(oidc.claims, "id_token.acr", {}) as {
+        essential?: boolean;
+        value?: string;
+      };
 
       if (!request || !request.essential || !request.value) {
         return Check.NO_NEED_TO_PROMPT;
@@ -25,7 +30,9 @@ loginPrompt.checks.add(
 
       return Check.REQUEST_PROMPT;
     },
-    ({ oidc }) => ({ acr: oidc.claims.id_token.acr }),
+    ({ oidc }) => ({
+      acr: oidc.claims?.["id_token"]?.["acr"],
+    }),
   ),
 );
